# perf: Better bit packing-unpacking algorithms