@mapleFU 2024-04-17
https://github.com/apache/arrow/issues/40845

I'm investigating improving the unpack in arrow, do you have some advices here?